Crate wasmer_engine_object_file

Object file backend for Wasmer compilers.

Given a compiler (such as CraneliftCompiler or LLVMCompiler) it generates a object file (.o file) and metadata which can be used to access it from other programming languages static.

Modules

artifact
Define ObjectFileArtifact to allow compiling and instantiating to be done as separate steps.

builder

engine

serialize

Structs

ObjectFile
The ObjectFile builder

ObjectFileArtifact
A compiled wasm module, ready to be instantiated.

ObjectFileEngine
A WebAssembly ObjectFile Engine.

Constants

VERSION
Version number of this crate.